Use sync.Cond for waiting on deps
This should be a lot cheaper than just blocking on channels, since we
know there's nothing left to do until the dep is finished.

I've also done a ReverseSorted on the dep graph so we start graph leaves
first, while should help with making forward progress (and waiting on
things that are more likely to have finished for status reporting).

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>